libc 2.38: backport upstream API patch
pg_query 2.2.1 cannot be compiled with glibc 2.38, because it redefines
`strchrnull`, see https://github.com/pganalyze/pg_query/issues/298

Instead of the patch provided in the above issue, we partially backport
the cleaner solution of 4f2eed2d589c5903916df3b189643e743149286e.
A difference with the 4.2.2 baseline is that the BSDs case was not present
in the `pg_config.h` file yet.